backend/requirements.txt
flask
flask-cors
python-dotenv
requests
scikit-learn
pandas
numpy
joblib
langchain
chromadb
sentence-transformers
transformers
huggingface-hub
